Add item uncheck behavior.
The default behavior for unchecking an item is to restore the item to it's previous position, before it was marked as checked. These previous item positions are stored in-memory and are based on the state of the checklist at that point in time, before it was marked as checked. With this in mind, the exact item restore position cannot be guaranteed.